unstable mode_t went to u32
Since this crate attempts to emulate UNIX style permissions on #[cfg(windows)] but the unstable std::os::linux::raw::mode_t type is only going to be  available on #[cfg(linux)], it broke
when mode_t was changed from i32 to u32.

Easy enough to fix... u32 and i32 are treated the same when you write them to the tar file, anyway. Not as easy to detect if mode_t changes, if you never compile for Linux though.

I couldn't get the tests to run without #![enable(deprecated)] because of os::platform::fs::MetadataExt on line 310, but that probably should be fixed outside of the scope of this. With deprecated enabled, the tests passed fine for me (and I run #[cfg(linux)]).