Split out a Builder from the Archive type
Using the same type for both building and reading is bound to lead to confusion
when the two are easily composable together, so add a separate and dedicated
type to just building archives. The `tar::Builder` type has the same API as
before, but now doesn't require calling the `finish()` function as it will
happen automatically on `Drop` or on `into_inner` (conveniently).