Preserve `\` characters in path names
Historically this library mapped all `\` characters into `/` characters to
handle encoding Windows paths. This is incorrect behavior, however, on Unix
where `\` is a valid character in filenames.

This commit tweaks the treatment of `\` to ignore it everywhere essentially.
Directory separators on Windows are translated to `/` as that's what the tar
specification indicates, which handles the translation on Windows naturally.