Don't fill atime and ctime (7-zip compat)
The atime and ctime fields in the GNU entry header seem to be
deprecated. GNU tar only fills them in when --incremental is passed and
the manpage calls that variation "old GNU-format".

Filling the two fields is problematic because 7-zip (which is rather
popular on Windows) interprets the timestamps as directories. WinRAR is
not affected, but libarchive ignores the two fields. Since they are not
well-supported, it's better to leave them empty.

Fixes: #70