Use the right name for @LongLink entries
These seem to be found with a name of `././@LongLink`, however
`copy_path_into` emits that as `@LongLink`. This is more tricky to fix
because `Path::components()` collapses that into `./@LongLink`. For now,
just special-case them.

The implementation here is not ideal since it will change the behavior
for non-GNU entries.