Move append_fs as method in Builder
Some other functions using append_fs are moved too.  The reason for this
change is to pass less parameter to append_fs.  Identification of hard
links will be implemented in a coming commit.  Without this change, one
more parameter of seen hard links will have to passed to append_fs.
That will be too messy.