fix #128: add context to _all_ error messages except `Read` and `Write`
Obviously I may have missed a few but I tried to be as complete as I
could be. I approached this by looking at the return type of _every_
function. If it was `io::Error` then I carefully steped through
the function making sure:
- the function _itself_ added context to errors
- OR the functions it was calling added context

In some cases context had to be added in both places, especially
the `to_octal` and `from_octal`. Those methods could tell _what_
was wrong (invalid input) but they couldn't say what the user
was _trying_ to do. Therefore I formatted their error and then
also added context at the caller's side.

Read and Write were not added since it is possible that the
callee passed in a Reader/Writer that provides better context
than we do. Since we do not know the _path_ we are reading/writing
from, we can't provide very good context anyway.