Issue #198: Safely overwrite files on extraction
Unlinking files before extraction is vulnerable to a TOCTOU race
condition, instead O_EXCL must be used when creating the file, and then
if that errors due to a conflicting file, it can be unlinked and the
operation retried. This has less syscalls in the common case, and one
more in the uncommon case of unpacking a tar over existing contents.