Permit choosing mode during extraction
This is important for reducing syscalls on cluster filesystems
where metadata opertions like chmod can be as slow as 2ms each.

Rather than requiring users to chmod post extraction permit
selection of mode during extraction of individual entries.

This will be complementary to ignore_permissions once mode setting
at creat(2) is in place, as then, for writable files at least, the
correct file mode can be created in a single syscall.

Possibly ignore_permissions will not be needed, if we can thread the
right permission logic through the process - I'm working on that
still.