** Add case for entry size in the pax extension **
This commit adds support for when the size of the entry does not reside
in the block header and instead with the associated local pax extension.
We check if there is a 'size' keyword present in the extension and use its
value as the entry size if the existing header size is zero.

Signed-off-by: Iris Liu <iris@onechronos.com>