Fix long-link paths with multibyte characters on Windows
Fixes an issue where when a truncated path was inserted into a header
after the long-link entry was already emitted the truncation happened on
a multibyte character boundary. The fix here is to always use
`str`-encoding for the header's path entry since it's not looked at
during exttraction on anyway and this keeps things compatible with Windows.

Closes #239