Release 0.4.41
There's only one notable change here, which is that
`Header::Deterministic` on Windows now uses the same timestamp
as is used on Unix.  If this breaks your use case, please
let us know in https://github.com/alexcrichton/tar-rs/pull/346

Besides that, there's a few testing, documentation, and internal
cleanups.  We're hoping to continue to merge PRs and address
issues; this release is small but is intending to maintain
momentum.

Signed-off-by: Colin Walters <walters@verbum.org>