Use zerocopy instead of raw transmute
This is a very old crate, and when I first started learning
Rust, I thought a tar parser was exactly the kind of thing
that should showcase Rust's claimed safety and speed.

I remember being surprised at the usage of `unsafe` here...thinking
it undermined some of the Rust claims.

I didn't think about it too much more though, but I now understand
that it wasn't exactly a language limitation (if we wanted to
we could just have `Header` and functions which reinterpret
fields manually) but nowadays the zerocopy crate has emerged
as the go-to default for ergonomically deriving and proving
at compile time the safety of these types of transmutations.

Let's start using it.

(Also use `#[derive(Default)]` instead of the unsafe `mem::zeroed()`
 for `GnuExtSparseHeader)

Now I also went ahead and added a `#[deny(unsafe_code)]` - there's
only a few calls to libc functions which could be replaced
with rustix or nix...but that's a different topic.

Adding a new dependency
-----------------------

I am aware that this crate is widely used, including by e.g. `cargo`.
I went ahead and proactively checked and it turns out that zerocopy
is already a dependency of `ahash`, which is already a cargo dependency.

Signed-off-by: Colin Walters <walters@verbum.org>