OSS-Fuzz: Use derive via Arbitrary feature flag (#416)
Before this change, the fuzzing target uses the arbitrary crate without its
derive feature flag, and instead imports both the Arbitrary type from the
arbitrary crate and the Arbitrary derive macro from the derive_arbitrary
crate.

When building in an environment that enables the `derive` feature of the
arbitrary crate, this causes an error due to the ambiguity produced by having
two separate imports of a derive macro called Arbitrary.

This change drops the separate derive_arbitrary crate and instead enables
the `derive` feature of the arbitrary crate, which resolves the ambiguity.